requirements.txt
Django
djangorestframework
django-cors-headers
django-webpack-loader